#2648: Transferred functionality from here from BinarizeMojo.java
